Fix zero column
Currently, `Goodcheck::Location` can have the column value `0`.
This is not useful for the output of the JSON reporter.

For example, if a text of `matched` is detected, I think that users expect:

- start_line: 2, start_column: 3
- end_line: 2, end_column: 9

```
1 |
2 | A matched text
3 |   ^~~~~~~
      3456789
```

The cause is that columns are calculated as a zero-started index.
This change aims to fix column values fitting to the expectation.